+++ sdk/src/events.rs
@@ -46,10 +46,10 @@ pub enum NexusEventKind {
     FoundingLeaderCapCreated(FoundingLeaderCapCreatedEvent),
     #[serde(rename = "GasSettlementUpdateEvent")]
     GasSettlementUpdate(GasSettlementUpdateEvent),
-    #[serde(rename = "PrekeyVaultCreatedEvent")]
-    PrekeyVaultCreated(PrekeyVaultCreatedEvent),
-    #[serde(rename = "PrekeyClaimedEvent")]
-    PrekeyClaimed(PrekeyClaimedEvent),
+    #[serde(rename = "PreKeyVaultCreatedEvent")]
+    PreKeyVaultCreated(PreKeyVaultCreatedEvent),
+    #[serde(rename = "PreKeyClaimedEvent")]
+    PreKeyClaimed(PreKeyClaimedEvent),
     // These events are unused for now.
     #[serde(rename = "ToolRegistryCreatedEvent")]
     ToolRegistryCreated(serde_json::Value),
@@ -210,23 +210,23 @@ pub struct GasSettlementUpdateEvent {
     pub was_settled: bool,
 }
 
-/// Fired by the Nexus Workflow when a new prekey vault is created. This happens
+/// Fired by the Nexus Workflow when a new pre_key vault is created. This happens
 /// on initial network setup.
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct PrekeyVaultCreatedEvent {
+pub struct PreKeyVaultCreatedEvent {
     pub vault: sui::ObjectID,
     pub crypto_cap: sui::ObjectID,
 }
 
-/// Fired by the Nexus Workflow when a prekey is claimed. Notifies about the
-/// remaining prekeys in the vault after the claim.
+/// Fired by the Nexus Workflow when a pre_key is claimed. Notifies about the
+/// remaining pre_keys in the vault after the claim.
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct PrekeyClaimedEvent {
-    /// What is the internal ID of the prekey that was claimed.
-    pub prekey_id: u32,
-    /// Which address claimed the prekey.
+pub struct PreKeyClaimedEvent {
+    /// What is the internal ID of the pre_key that was claimed.
+    pub pre_key_id: u32,
+    /// Which address claimed the pre_key.
     pub claimed_by: sui::ObjectID,
-    /// How many prekeys are left in the vault after this claim.
+    /// How many pre_keys are left in the vault after this claim.
     #[serde(
         deserialize_with = "deserialize_sui_u64",
         serialize_with = "serialize_sui_u64"